move stuff to partials
