+++ Content.Server/Chemistry/Components/ChemMasterComponent.cs
@@ -37,12 +37,5 @@ namespace Content.Server.Chemistry.Components
         /// </summary>
         [DataField]
         public ChemMasterDrawSource DrawSource = ChemMasterDrawSource.External; ///Starlight edit
-
-        /// <summary>
-        ///     TRIESTE SPECIFIC
-        ///     The transfer amount so the UI buttons make a click.
-        /// </summary>
-        [DataField]
-        public ChemMasterReagentAmount TransferAmount = ChemMasterReagentAmount.U5;
     }
 }
